test: Update test codes and refactor out registry_ctx
* Temporarily remove "timeout" example session as it interfers with
  the DRF/LIFO scheduler test cases.  We should use a separate example
  session list.
